extracting KeysPipe into new SampleUtilitiesModule
