+++ src/app/sample-utilities/sample-utilities.module.ts
@@ -1,0 +1,18 @@
+import { ModuleWithProviders, NgModule } from "@angular/core";
+import { KeysPipe } from "./keys-pipe/keys-pipe";
+
+@NgModule({
+    declarations: [
+        KeysPipe
+    ],
+    exports: [
+        KeysPipe
+    ]
+})
+export class SamplesUtilitiesModule {
+    public static forRoot(): ModuleWithProviders {
+        return {
+            ngModule: SamplesUtilitiesModule
+        };
+    }
+}
